+++ code/packages/perl/heap/t/00-load.t
@@ -1,0 +1,10 @@
+use strict;
+use warnings;
+use Test2::V0;
+
+ok lives {
+    require CodingAdventures::Heap;
+    1;
+}, 'CodingAdventures::Heap loads';
+
+done_testing;